+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added user setting: number formatting
 - Waterfall: highlight numbers in comments
+- Waterfall: text wrapping on the x-axis
 
 ### Changed